separate DDict logic into its own module
created zstd_ddict.c within lib/decompress
